pyproject.toml
[project]
authors = [{ name = "Jermiah Joseph", email = "jermiahjoseph98@gmail.com" }]
dependencies = [
    "torch>=1.10.0", 
    "torchvision>=0.11.0", 
    "matplotlib>=3.4.0", 
    "click>=8.0.0", 
    "tqdm>=4.62.0", 
    "numpy>=1.20.0", 
    "pillow>=8.2.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.0.0",
    "rich>=10.0.0",
]
name = "simple-ml-project"
description = "A robust ML project for MNIST classification"
readme = "README.md"
requires-python = ">= 3.11"
version = "0.1.0"

[project.scripts]
sml = "simple_ml_project.cli.commands:cli"
sml-train = "simple_ml_project.cli.commands:train_model"
sml-predict = "simple_ml_project.cli.commands:predict"
sml-dataset = "simple_ml_project.cli.commands:dataset"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py311"]
